requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.54.0
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
